Added preinvoke logic
